client/requirements.txt
pyperclip==1.8.2
pillow==10.4.0
requests==2.32.3
colorama==0.4.6